fastapi/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
autopep8==2.0.4
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
jwt==1.3.1
#mariadb==1.1.10
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
mysqlclient==2.2.4
orjson==3.9.10
packaging==23.2
passlib==1.7.4
protobuf==4.21.12
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
setuptools==69.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.42.0
